Refactor Name class into a separate package to avoid code duplication
* Both Property and Appointment classes can use the same Name class
