+++ src/test/java/seedu/address/logic/parser/AddPropertyCommandParserTest.java
@@ -24,9 +24,9 @@ import static seedu.address.testutil.TypicalProperties.MAYFAIR;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.AddPropertyCommand;
+import seedu.address.model.name.Name;
 import seedu.address.model.property.Address;
 import seedu.address.model.property.Deadline;
-import seedu.address.model.property.Name;
 import seedu.address.model.property.PostalCode;
 import seedu.address.model.property.Property;
 import seedu.address.model.property.Type;
